Create button and modal for New Book and apply JS
Modal is created by using the dialog element.
Applied a 'click' listener on the New Book button. This will prompt the
modal to open/show.
Upon clicking the submit button, a function will be triggered to do the
following: prevent form from sending data to the server, make a Book
object by using user's input and add it to myLibrary array, and finally,
close the modal.